[fakeintake] move redact to dedicated file
