+++ test/fakeintake/server/server_test.go
@@ -654,32 +654,3 @@ func PostSomeRealisticLogs(t *testing.T, url string) {
 	require.NoError(t, err, "Error on POST request")
 	defer response.Body.Close()
 }
-
-func TestRedactHeader(t *testing.T) {
-	header := http.Header{
-		"Foo":       []string{"bar"},
-		"Key":       []string{"AAAAAAA"},
-		"appkEy":    []string{"BBBBBBB"},
-		"other-KeY": []string{"CCCCCCC", "DDDDDDD"},
-	}
-	redactedHeader := redactHeader(header)
-	expectedHeader := http.Header{
-		"Foo":       []string{"bar"},
-		"Key":       []string{"<redacted>"},
-		"Appkey":    []string{"<redacted>"},
-		"Other-Key": []string{"<redacted>"},
-	}
-	expectedKeys := []string{}
-	for key, expectedValues := range expectedHeader {
-		redactedValues := redactedHeader[key]
-		assert.Equal(t, expectedValues, redactedValues, "unexpected value at key %s", key)
-		expectedKeys = append(expectedKeys, key)
-	}
-	redactedKeys := []string{}
-	for key, redactedValues := range redactedHeader {
-		expectedValues := expectedHeader[key]
-		assert.Equal(t, expectedValues, redactedValues, "unexpected value at key %s", key)
-		redactedKeys = append(redactedKeys, key)
-	}
-	assert.Equal(t, len(expectedKeys), len(redactedKeys), "unexpected length in keys")
-}
